Add a rating to different kombucha recipes
Make a Rating's model with the following rules:

should have a score that returns a number between 1 and 5.
Users can only have 1 rating per kombucha.
Add API endpoints so that users can:

See their rating
Create a rating
Update an existing rating
In each endpoint, add an attribute in the response to display the average rating of a given kombucha tea. This average rating should be based on all user ratings for that particular kombucha tea.