Add Advanced Filtering
Included Ingredients: return kombucha teas that contain the ingredients passed to it. Example: find all kombucha's that have coffee in them.
Not Included Ingredients: return kombucha teas that do not contain the ingredients passed to the endpoint. Example: I don't want recipes that contain coffee
Popular recipes: return kombucha teas that are in 'x' percent or greater of Kombucha flights.